WIP, About me section being built, file relocations,
