+++ src/index.css
@@ -9,7 +9,7 @@
 
   color-scheme: light dark;
   color: rgba(255, 255, 255, 0.87);
-  background-color: #242424;
+  background-color: #141418;
 
   font-synthesis: none;
   text-rendering: optimizeLegibility;
