+++ src/pages/Home/Home.tsx
@@ -1,3 +1,5 @@
+import { AboutMeHomeCard } from "../../Components/AboutMeHomeCard/AboutMeHomeCard";
+
 export default function Home() {
-  return <></>;
+  return <AboutMeHomeCard />;
 }
